Moving rate function to separate RatingHelper class.
